info30005/requirements.txt
Django==1.5
PyYAML==3.10
South==0.7.6
dj-database-url==0.2.1
gunicorn==0.17.4
matplotlib==1.1.1
nltk==2.0.1
node==0.9.11
numpy==1.6.2
odict==1.5.1
plumber==1.2
psycopg2==2.5
recaptcha-client==1.0.6
six==1.3.0
virtualenv==1.9.1
wsgiref==0.1.2
zope.component==4.1.0
zope.deprecation==4.0.2
zope.event==4.0.2
zope.interface==4.0.5
zope.lifecycleevent==4.0.2
